requirements.txt.txt
python-telegram-bot==21.4
requests
beautifulsoup4
